python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "renderscope"
version = "0.1.0"
description = "CLI tool and library for benchmarking, comparing, and cataloging open-source rendering engines"
readme = "README.md"
license = { text = "Apache-2.0" }
authors = [
    { name = "Ashutosh Mishra" },
]
requires-python = ">=3.10"
keywords = [
    "rendering",
    "benchmarking",
    "computer graphics",
    "comparison",
    "path tracing",
    "ray tracing",
]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Multimedia :: Graphics :: 3D Rendering",
    "Topic :: Scientific/Engineering :: Visualization",
    "Typing :: Typed",
]

dependencies = [
    "typer>=0.9.0",
    "rich>=13.0",
    "shellingham>=1.5",
    "pydantic>=2.0",
    "psutil>=5.9",
    "numpy>=1.24",
    "Pillow>=10.0",
    "scikit-image>=0.21",
]

[project.optional-dependencies]
ml = [
    "torch>=2.0",
    "torchmetrics>=1.0",
]
plots = [
    "matplotlib>=3.7",
]
cv = [
    "opencv-python-headless>=4.8",
]
all = [
    "renderscope[ml,plots,cv]",
]
dev = [
    "pytest>=8",
    "pytest-cov>=5.0",
    "mypy>=1.10",
    "ruff>=0.4",
]

[project.scripts]
renderscope = "renderscope.cli.main:app"

[project.urls]
Homepage = "https://github.com/renderscope-dev/renderscope"
Repository = "https://github.com/renderscope-dev/renderscope"
Documentation = "https://github.com/renderscope-dev/renderscope#readme"
Issues = "https://github.com/renderscope-dev/renderscope/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/renderscope"]

[tool.hatch.build.targets.wheel.force-include]
"../data/renderers" = "renderscope/data/renderers"

[tool.ruff]
target-version = "py310"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort (import sorting)
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "TCH",  # flake8-type-checking
    "RUF",  # ruff-specific rules
]
ignore = [
    "E501",  # line-too-long (handled by formatter)
]

[tool.ruff.lint.per-file-ignores]
# Typer uses function calls in argument defaults — this is the standard pattern.
# Path must remain a runtime import because Typer resolves type hints at runtime.
"src/renderscope/cli/*.py" = ["B008", "TC003"]
# Pydantic needs cross-model imports at runtime for field type resolution.
"src/renderscope/models/benchmark.py" = ["TC001"]
# Test fixtures use Path in annotations resolved by pytest — keep as runtime import.
"tests/*.py" = ["TC003"]

[tool.ruff.lint.isort]
known-first-party = ["renderscope"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = ["psutil", "psutil.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["torch", "torch.*", "torchmetrics", "torchmetrics.*", "matplotlib", "matplotlib.*", "cv2", "cv2.*", "mitsuba", "mitsuba.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["skimage", "skimage.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["imageio", "imageio.*", "OpenEXR", "Imath"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["pyluxcore", "pyluxcore.*", "pyfilament", "pyfilament.*", "ospray", "ospray.*"]
ignore_missing_imports = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
